refactor: Move image_processing nodes to individual files
Implements ML-597
